+++ app/graphql/types/subscriptions/transaction.rb
@@ -6,7 +6,7 @@ module Types
       graphql_name 'SubscriptionsTransaction'
 
       field :id, ID, null: false
-      field :amount, Integer, null: false
+      field :amount, Float, null: false
       field :currency, Types::Common::Currency, null: false
       field :invoice_web_url, String, null: false
       field :invoice_pdf_url, String, null: false
@@ -16,7 +16,7 @@ module Types
       field :period_end_at, GraphQL::Types::ISO8601Date, null: false
       field :discount_name, String, null: true
       field :discount_percentage, Float, null: true
-      field :discount_amount, Integer, null: true
+      field :discount_amount, Float, null: true
       field :discount_id, String, null: true
     end
   end
